moving tests, adding comments to regex
