+++ packages/cli/commands/validate.js
@@ -1,15 +1,10 @@
-const {
-  addConfigOptions,
-  addAccountOptions,
-  addOverwriteOptions,
-} = require('../lib/commonOpts');
-const validateTheme = require('./validation/validateTheme');
+const { addConfigOptions, addAccountOptions } = require('../lib/commonOpts');
+const validateTheme = require('./validate/validateTheme');
 
 exports.command = 'validate';
 exports.describe = 'Commands for working with validation';
 
 exports.builder = yargs => {
-  addOverwriteOptions(yargs, true);
   addConfigOptions(yargs, true);
   addAccountOptions(yargs, true);
