+++ packages/cli/commands/validate/validateTheme.js
