+++ packages/cli/lib/validators/__tests__/themeConfigValdiator.js
